+++ client/src/modules/journal/journal.js
@@ -6,7 +6,8 @@ JournalController.$inject = [
   'GridFilteringService', 'GridColumnService', 'JournalConfigService',
   'SessionService', 'NotifyService', 'TransactionService', 'GridEditorService',
   'bhConstants', '$state', 'uiGridConstants', 'ModalService', 'LanguageService',
-  'AppCache', 'Store', 'uiGridGroupingConstants',
+  'AppCache', 'Store', 'uiGridGroupingConstants', 'ExportService', 'FindEntityService',
+  'FilterService', '$rootScope', '$filter', '$translate'
 ];
 
 /**
@@ -28,11 +29,37 @@ JournalController.$inject = [
  *   - Tun trial balance validation on transactions
  *
  * @todo Propose utility bar view design
- *
- * @module bhima/controllers/JournalController
  */
-function JournalController(Journal, Sorting, Grouping, Filtering, Columns, Config, Session, Notify, Transactions, Editors, bhConstants, $state, uiGridConstants, Modal, Languages, AppCache, Store, uiGridGroupingConstants) {
+function JournalController(Journal, Sorting, Grouping,
+  Filtering, Columns, Config, Session, Notify, Transactions, Editors,
+  bhConstants, $state, uiGridConstants, Modal, Languages, AppCache, Store,
+  uiGridGroupingConstants, Export, FindEntity, Filters, $rootScope, $filter, $translate) {
+
+  // Journal utilities
+  var sorting;
+  var grouping;
+  var filtering;
+  var columnConfig;
+  var transactions;
+
+  var filter = new Filters();
+
+  /** @const the cache alias for this controller */
+  var cacheKey = 'Journal';
+
+  // top level cache
+  var cache = AppCache(cacheKey + '-module');
   var vm = this;
+  vm.filter = filter;
+
+  // number of all of the transactions in the system
+  Journal.count()
+    .then(function (data) {
+      vm.numberTotalSystemTransactions = data[0].number_transactions;
+    })
+    .catch(function (error) {
+      Notify.handleError(error);
+    });
 
   /** @constants */
   vm.ROW_EDIT_FLAG = bhConstants.transactions.ROW_EDIT_FLAG;
@@ -42,32 +69,18 @@ function JournalController(Journal, Sorting, Grouping, Filtering, Columns, Confi
   // @todo - this doesn't work with the ui-grid-datepicker-edit library yet.
   vm.DATEPICKER_OPTIONS = { format: bhConstants.dates.format };
 
-  // Journal utilities
-  var sorting;
-  var grouping;
-  var filtering;
-  var columnConfig;
-  var transactions;
-  var editors; // editors is affected but not used
-
-  /** @const the cache alias for this controller */
-  var cacheKey = 'Journal';
-
-  // filter cache
-  var cache = AppCache(cacheKey + '-filters');
-
-  var vm = this;
-
   vm.enterprise = Session.enterprise;
 
   // gridOptions is bound to the UI Grid and used to configure many of the
   // options, it is also used by the grid to expose the API
   vm.gridOptions = {
-    enableColumnMenus : false,
-    showColumnFooter  : true,
-    appScopeProvider  : vm,
-    flatEntityAccess  : true,
-    rowTemplate       : '/modules/templates/grid/transaction.row.html',
+    enableColumnMenus          : false,
+    showColumnFooter           : true,
+    appScopeProvider           : vm,
+    flatEntityAccess           : true,
+    enableGroupHeaderSelection : true,
+    enableRowHeaderSelection   : true,
+    rowTemplate                : '/modules/templates/grid/transaction.row.html'
   };
 
   vm.grouped = angular.isDefined(cache.grouped) ? cache.grouped : false;
@@ -76,10 +89,9 @@ function JournalController(Journal, Sorting, Grouping, Filtering, Columns, Confi
   // configuration options
   sorting = new Sorting(vm.gridOptions);
   filtering = new Filtering(vm.gridOptions, cacheKey);
-  grouping  = new Grouping(vm.gridOptions, true, 'trans_id', vm.grouped, false);
+  grouping = new Grouping(vm.gridOptions, true, 'trans_id', vm.grouped, false);
   columnConfig = new Columns(vm.gridOptions, cacheKey);
   transactions = new Transactions(vm.gridOptions);
-  editors = new Editors(vm.gridOptions);
 
   // attaching the filtering object to the view
   vm.filtering = filtering;
@@ -91,7 +103,6 @@ function JournalController(Journal, Sorting, Grouping, Filtering, Columns, Confi
   vm.transactions = transactions;
 
   vm.onRemoveFilter = onRemoveFilter;
-  vm.clearFilters = clearFilters;
 
   vm.cancelEdit = cancelEdit;
 
@@ -121,58 +132,130 @@ function JournalController(Journal, Sorting, Grouping, Filtering, Columns, Confi
    *      other columns. This can be avoided by setting default sort and group.
    */
   var columns = [
-    { field: 'uuid', displayName : 'TABLE.COLUMNS.ID', headerCellFilter: 'translate', visible: false, enableCellEdit: false },
-    { field: 'project_name', displayName : 'TABLE.COLUMNS.PROJECT', headerCellFilter: 'translate', visible: false, enableCellEdit: false },
-    { field: 'period_end', displayName : 'TABLE.COLUMNS.PERIOD', headerCellFilter: 'translate' , cellTemplate : 'modules/templates/bhPeriod.tmpl.html', visible : false, enableCellEdit : false},
-    { field: 'trans_id',
-      displayName : 'TABLE.COLUMNS.TRANSACTION',
-      headerCellFilter: 'translate',
+    { field            : 'uuid',
+      displayName      : 'TABLE.COLUMNS.ID',
+      headerCellFilter : 'translate',
+      visible          : false,
+      enableCellEdit   : false },
+
+    { field            : 'project_name',
+      displayName      : 'TABLE.COLUMNS.PROJECT',
+      headerCellFilter : 'translate',
+      visible          : false,
+      enableCellEdit   : false },
+
+    { field            : 'period_end',
+      displayName      : 'TABLE.COLUMNS.PERIOD',
+      headerCellFilter : 'translate' ,
+      cellTemplate     : 'modules/templates/bhPeriod.tmpl.html',
+      visible          : false,
+      enableCellEdit   : false },
+
+    { field            : 'trans_id',
+      displayName      : 'TABLE.COLUMNS.TRANSACTION',
+      headerCellFilter : 'translate',
       sortingAlgorithm : sorting.transactionIds,
-      // sort : { priority : 0, direction : 'asc' },
-      enableCellEdit: false,
-      width : 110,
-      cellTemplate : 'modules/journal/templates/hide-groups-label.cell.html',
+      enableCellEdit   : false,
+      width            : 110,
+      cellTemplate     : 'modules/journal/templates/hide-groups-label.cell.html' },
+
+    { field                            : 'trans_date',
+      displayName                      : 'TABLE.COLUMNS.DATE',
+      headerCellFilter                 : 'translate',
+      cellFilter                       : 'date:"' + bhConstants.dates.format + '"',
+      filter                           : { condition: filtering.filterByDate },
+      editableCellTemplate             : 'modules/journal/templates/date.edit.html',
+      treeAggregationType              : uiGridGroupingConstants.aggregation.MIN,
+      customTreeAggregationFinalizerFn : function (aggregation) {
+        aggregation.rendered = $filter('date')(aggregation.value, bhConstants.dates.format);
+      },
+      enableCellEdit     : true,
+      footerCellTemplate : '<i></i>' },
+
+    { field                : 'hrRecord',
+      displayName          : 'TABLE.COLUMNS.RECORD',
+      headerCellFilter     : 'translate',
+      visible              : true,
+      treeAggregationType  : uiGridGroupingConstants.aggregation.MIN,
+      treeAggregationLabel : '',
+      enableCellEdit       : false,
+      footerCellTemplate   : '<i></i>' },
+
+    { field              : 'description',
+      displayName        : 'TABLE.COLUMNS.DESCRIPTION',
+      headerCellFilter   : 'translate',
+      footerCellTemplate : '<i></i>' },
+
+    { field                : 'account_number',
+      displayName          : 'TABLE.COLUMNS.ACCOUNT',
+      editableCellTemplate : '<div><form name="inputForm"><div ui-grid-edit-account></div></form></div>',
+      enableCellEdit       : true,
+      headerCellFilter     : 'translate',
     }, {
-      field : 'trans_date',
-      displayName : 'TABLE.COLUMNS.DATE',
-      headerCellFilter: 'translate',
-      cellFilter : 'date:"' + bhConstants.dates.format + '"',
-      filter : { condition : filtering.filterByDate },
-      editableCellTemplate: 'modules/journal/templates/date.edit.html',
-      enableCellEdit: true,
-      footerCellTemplate:'<i></i>',
-    },
-    { field : 'hrRecord', displayName : 'TABLE.COLUMNS.RECORD', headerCellFilter: 'translate', visible: true, enableCellEdit : false },
-    { field : 'description', displayName : 'TABLE.COLUMNS.DESCRIPTION', headerCellFilter: 'translate', footerCellTemplate:'<i></i>' },
-    { field : 'account_number', displayName : 'TABLE.COLUMNS.ACCOUNT', headerCellFilter: 'translate' },
-    { field : 'debit_equiv',
-      displayName : 'TABLE.COLUMNS.DEBIT',
-      headerCellFilter: 'translate',
-      treeAggregationType : uiGridGroupingConstants.aggregation.SUM,
+      field                            : 'debit_equiv',
+      displayName                      : 'TABLE.COLUMNS.DEBIT',
+      headerCellFilter                 : 'translate',
+      treeAggregationType              : uiGridGroupingConstants.aggregation.SUM,
       customTreeAggregationFinalizerFn : function (aggregation) {
         aggregation.rendered = aggregation.value;
       },
-      enableFiltering: false
-    }, {
-      field : 'credit_equiv',
-      displayName : 'TABLE.COLUMNS.CREDIT',
-      headerCellFilter: 'translate',
-      treeAggregationType : uiGridGroupingConstants.aggregation.SUM,
+      enableFiltering : false },
+
+    { field                            : 'credit_equiv',
+      displayName                      : 'TABLE.COLUMNS.CREDIT',
+      headerCellFilter                 : 'translate',
+      treeAggregationType              : uiGridGroupingConstants.aggregation.SUM,
       customTreeAggregationFinalizerFn : function (aggregation) {
         aggregation.rendered = aggregation.value;
       },
-      enableFiltering: false
+      enableFiltering : false },
+
+    { field            : 'currencyName',
+      displayName      : 'TABLE.COLUMNS.CURRENCY',
+      headerCellFilter : 'translate',
+      visible          : false,
+      enableCellEdit   : false },
+
+    { field            : 'debit',
+      displayName      : 'TABLE.COLUMNS.DEBIT_SOURCE',
+      headerCellFilter : 'translate',
+      visible          : false,
+      cellTemplate     : '/modules/journal/templates/debit.grid.html',
+      enableCellEdit   : false },
+
+    { field            : 'credit',
+      displayName      : 'TABLE.COLUMNS.CREDIT_SOURCE',
+      headerCellFilter : 'translate',
+      visible          : false,
+      cellTemplate     : '/modules/journal/templates/credit.grid.html',
+      enableCellEdit   : false },
+
+    { field                : 'hrEntity',
+      displayName          : 'TABLE.COLUMNS.RECIPIENT',
+      headerCellFilter     : 'translate',
+      editableCellTemplate : '/modules/journal/templates/entity.edit.html',
+      visible              : true },
+
+    { field            : 'hrReference',
+      displayName      : 'TABLE.COLUMNS.REFERENCE',
+      headerCellFilter : 'translate',
+      visible          : true },
+
+    { field            : 'display_name',
+      displayName      : 'TABLE.COLUMNS.RESPONSIBLE',
+      headerCellFilter : 'translate',
+      visible          : false,
+      enableCellEdit   : false },
+
+    { field            : 'actions',
+      displayName      : '',
+      headerCellFilter : 'translate',
+      visible          : true,
+      enableCellEdit   : false,
+      cellTemplate     : '/modules/journal/templates/actions.cell.html',
+      allowCellFocus   : false,
+      enableFiltering  : false,
     },
-    { field : 'currencyName', displayName : 'TABLE.COLUMNS.CURRENCY', headerCellFilter: 'translate', visible: false, enableCellEdit: false },
-    { field : 'hrEntity', displayName : 'TABLE.COLUMNS.RECIPIENT', headerCellFilter: 'translate', visible: true },
-    { field : 'hrReference', displayName : 'TABLE.COLUMNS.REFERENCE', headerCellFilter: 'translate', visible: true },
-    { field : 'display_name', displayName : 'TABLE.COLUMNS.RESPONSIBLE', headerCellFilter: 'translate', visible: false, enableCellEdit: false },
-    { field : 'actions', displayName : '', headerCellFilter: 'translate',
-      visible: true, enableCellEdit: false,
-      cellTemplate: '/modules/journal/templates/actions.cell.html',
-      allowCellFocus: false,
-      enableFiltering: false,
-    }
   ];
 
   vm.gridOptions.columnDefs = columns;
@@ -182,10 +265,12 @@ function JournalController(Journal, Sorting, Grouping, Filtering, Columns, Confi
     columnConfig.openConfigurationModal();
   };
 
-  //This function opens a modal, to let the user posting transaction to the general ledger
+  // This function opens a modal, to let the user posting transaction to the general ledger
   vm.openTrialBalanceModal = function openTrialBalanceModal() {
+    var numberSelectedGroup = vm.grouping.getSelectedGroups().length;
+
     // make sure a row is selected before running the trial balance
-    if (grouping.selectedRowCount < 1) {
+    if (numberSelectedGroup === 0) {
       Notify.warn('POSTING_JOURNAL.WARNINGS.NO_TRANSACTIONS_SELECTED');
       return;
     }
@@ -193,8 +278,8 @@ function JournalController(Journal, Sorting, Grouping, Filtering, Columns, Confi
     $state.go('trialBalanceMain', { records: vm.grouping.getSelectedGroups() });
   };
 
-  // display the journal printable report of selected transactions
-  vm.openJournalReport = function openJournalReport() {
+  // format Export Parameters
+  function formatExportParameters(type) {
     // make sure a row is selected before running the trial balance
     if (grouping.selectedRowCount < 1) {
       Notify.warn('POSTING_JOURNAL.WARNINGS.NO_TRANSACTIONS_SELECTED');
@@ -205,10 +290,27 @@ function JournalController(Journal, Sorting, Grouping, Filtering, Columns, Confi
       return trans.uuid;
     });
 
+    return { renderer: type || 'pdf', lang: Languages.key, uuids: uuids };
+  }
+
+  // display the journal printable report of selected transactions
+  vm.openJournalReport = function openJournalReport() {
     var url = '/reports/finance/journal';
-    var params = { renderer: 'pdf', lang: Languages.key, uuids: uuids };
-    Modal.openReports({ url: url, params: params })
-      .catch(angular.noop);
+    var params = formatExportParameters('pdf');
+
+    if (!params) { return; }
+
+    Modal.openReports({ url: url, params: params });
+  };
+
+  // export data into csv file
+  vm.exportFile = function exportFile() {
+    var url = '/reports/finance/journal';
+    var params = formatExportParameters('csv');
+
+    if (!params) { return; }
+
+    Export.download(url, params, 'POSTING_JOURNAL.TITLE');
   };
 
   function errorHandler(error) {
@@ -220,14 +322,23 @@ function JournalController(Journal, Sorting, Grouping, Filtering, Columns, Confi
   function load(options) {
     vm.loading = true;
     vm.hasError = false;
+    vm.gridOptions.gridFooterTemplate = null;
+    vm.gridOptions.showGridFooter = false;
+
+    // number of transactions downloaded and shown in the current journal
+    var numberCurrentGridTransactions = 0;
 
     // @fixme
     Journal.grid(null, options)
       .then(function (records) {
+        // To Get the number of transaction
+        numberCurrentGridTransactions =  records.aggregate.length;
 
         // pre process data - this should be done in a more generic way in a service
         vm.gridOptions.data = transactions.preprocessJournalData(records);
-
+        vm.gridOptions.showGridFooter = true;
+        vm.gridOptions.gridFooterTemplate = '<div><strong>' + $translate.instant('FORM.INFO.NUM_TRANSACTION') +
+          ' : ' + numberCurrentGridTransactions + ' / ' + vm.numberTotalSystemTransactions + '</strong></div>';
         transactions.applyEdits();
 
         // try to unfold groups
@@ -255,48 +366,39 @@ function JournalController(Journal, Sorting, Grouping, Filtering, Columns, Confi
 
   // open search modal
   vm.openSearchModal = function openSearchModal() {
-    var parameters = angular.copy(vm.filters);
-    Config.openSearchModal(parameters)
-      .then(function (options) {
-        // if the options are not returned or have not changed, do not refresh
-        // the data source
-        if (angular.equals(options, vm.filters)) { return; }
+    var filtersSnapshot = Journal.filters.formatHTTP();
 
-        // bind filters to the view and format appropriate
-        cacheFilters(options);
+    Config.openSearchModal(filtersSnapshot)
+      .then(function (changes) {
 
-        // turn loading on
+        Journal.filters.replaceFilters(changes);
+
+        Journal.cacheFilters();
+        vm.latestViewFilters = Journal.filters.formatView();
+
         toggleLoadingIndicator();
-
-        return load(options);
+        return load(Journal.filters.formatHTTP(true));
       })
       .catch(angular.noop);
   };
 
-  // save the parameters to use later.  Formats the parameters in filtersFmt for the filter toolbar.
-  function cacheFilters(filters) {
-    vm.filters = cache.filters = filters;
-    vm.filtersFmt = Journal.formatFilterParameters(filters);
-    vm.filterBarHeight = (vm.filtersFmt.length > 0) ?
-      bhConstants.utilBar.expandedHeightStyle : bhConstants.utilBar.collapsedHeightStyle;
-  }
-
   // remove a filter with from the filter object, save the filters and reload
   function onRemoveFilter(key) {
-    delete vm.filters[key];
-    cacheFilters(vm.filters);
-    load(vm.filters);
-  }
+    Journal.removeFilter(key);
 
-  // clears the filters by forcing a cache of an empty array
-  function clearFilters() {
-    cacheFilters({});
-    load({});
+    Journal.cacheFilters();
+    vm.latestViewFilters = Journal.filters.formatView();
+
+    return load(Journal.filters.formatHTTP(true));
   }
 
   vm.editTransaction = editTransaction;
   function editTransaction(row) {
-    vm.filterBarHeight = bhConstants.utilBar.expandedHeightStyle;
+    vm.filterBarHeight = bhConstants.utilBar.journalHeightStyle;
+
+    // expand the row
+    vm.grouping.unfoldGroup(row);
+
     transactions.edit(row);
 
     // disable inline filtering when editing
@@ -313,7 +415,7 @@ function JournalController(Journal, Sorting, Grouping, Filtering, Columns, Confi
         load(vm.filters);
       })
       .catch(function (error) {
-        if(error === 'POSTING_JOURNAL.ERRORS.UNBALANCED_TRANSACTIONS'){
+        if (!error.status) {
           Notify.warn(error);
         } else {
           Notify.handleError(error);
@@ -332,21 +434,55 @@ function JournalController(Journal, Sorting, Grouping, Filtering, Columns, Confi
     }
   };
 
+  vm.saveAccountEdit = function saveAcconutEdit(row, account) {
+    row.account_id = account.id;
+    row.account_name = account.hrlabel;
+    $rootScope.$emit(uiGridEditConstants.events.END_CELL_EDIT);
+  };
+
   function cancelEdit() {
+    vm.filterBarHeight = bhConstants.utilBar.collapsedHeightStyle;
+
     // @TODO this should return a promise in a uniform standard with `saveTransaction`
     transactions.cancel();
 
     // ensure data that has been changed is up to date from the server
     // remove any additional or temporary rows
-    load(vm.filters);
+    load(Journal.filters.formatHTTP(true));
   }
 
   // runs on startup
   function startup() {
-    vm.filters = cache.filters;
-    vm.filtersFmt = Journal.formatFilterParameters(cache.filters || {});
-    load(vm.filters);
+    load(Journal.filters.formatHTTP(true));
+    vm.latestViewFilters = Journal.filters.formatView();
   }
 
+  // ===================== edit entity ===============================
+
+  // expose to the view
+  vm.openEntityModal = openEntityModal;
+  vm.removeEntity = removeEntity;
+
+  // open find entity modal
+  function openEntityModal(row) {
+    FindEntity.openModal()
+      .then(function (entity) {
+        if (!entity) { return; }
+
+        row.hrEntity = entity.hrEntity;
+        transactions.editCell(row, 'entity_uuid', entity.uuid);
+      });
+  }
+
+  // remove the entity
+  function removeEntity(row) {
+    if (!row.hrEntity) { return; }
+
+    transactions.editCell(row, 'entity_uuid', null);
+    delete row.entity_uuid;
+    delete row.hrEntity;
+  }
+
+  // ===================== end edit entity ===========================
   startup();
 }